+++ src/main/java/leets/attendance/repository/AttendanceRepository.java
@@ -1,7 +1,7 @@
 package leets.attendance.repository;
 
-import leets.attendance.entity.Attendances;
-import leets.attendance.entity.User;
+import leets.attendance.domain.Attendances;
+import leets.attendance.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.time.LocalDate;
@@ -10,6 +10,5 @@ import java.util.List;
 public interface AttendanceRepository extends JpaRepository<Attendances, Long> {
     List<Attendances> findByUserAndDate(User user, LocalDate date);
     List<Attendances> findByUser(User user);
-    List<Attendances> findByUserAndDateBefore(User user, LocalDate date);
     List<Attendances> findAllByUser(User user);
 }
