+++ src/main/java/leets/attendance/service/AuthService.java
@@ -3,12 +3,12 @@ package leets.attendance.service;
 import jakarta.transaction.Transactional;
 import leets.attendance.dto.UserResponseDTO;
 import leets.attendance.dto.TokenDTO;
-import leets.attendance.entity.Attendances;
-import leets.attendance.entity.User;
-import leets.attendance.entity.WeekEnum;
+import leets.attendance.domain.Attendances;
+import leets.attendance.domain.User;
+import leets.attendance.domain.WeekEnum;
 import leets.attendance.repository.AttendanceRepository;
 import leets.attendance.repository.UserRepository;
-import leets.attendance.util.TokenProvider;
+import leets.attendance.domain.TokenProvider;
 import leets.attendance.dto.UserRequestDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
